go.mod: Add a bugfix
One of the changes when switching from `go get` to `go mod` is that the
default version of git dependencies changes.  With `go get` the default is
git HEAD.  With `go mod` it's the most recent tagged release (if any exist,
otherwise it falls back to HEAD).

contrib.go.opencensus.io/exporter/ocagent 0.4.0 came out recently.
Unfortunately it won't build--it relies on features in go.opencensus.io
that are in "master", but aren't in the most recent release (0.18.0).  On
ocagent "master", they've fixed their go.mod file to depend on a specific
commit of go.opencensus.io, but haven't released that fix as 0.4.1.  So
specifically depend on that commit, so we don't inherit build breakage.